Fix imports and move util functions to util.py
