+++ tests/odwr/test_upstream.py
@@ -32,18 +32,19 @@ from datetime import datetime, timedelta, timezone
 import pytest
 import requests
 
-from userCode.odwr.tests.lib import (
-    assert_date_in_range,
-    now_as_oregon_datetime,
-)
+
 from userCode.odwr.lib import (
     download_oregon_tsv,
-    from_oregon_datetime,
     generate_oregon_tsv_url,
     parse_oregon_tsv,
-    to_oregon_datetime,
 )
 from userCode.odwr.types import START_OF_DATA
+from userCode.util import (
+    assert_date_in_range,
+    now_as_oregon_datetime,
+    from_oregon_datetime,
+    to_oregon_datetime
+)
 
 
 @pytest.mark.parametrize(
